move property validators to their own folder
